:truck: Simplify CLI app entrypoint
