+++ ice_station_zebra/cli/__init__.py
@@ -1,0 +1,6 @@
+from .main import app, hydra_override
+
+__all__ = [
+    "app",
+    "hydra_override",
+]
